chore: export `ModuleBasicManager` for reuse by integrators
