+++ cmd/nobled/main.go
@@ -2,11 +2,38 @@ package main
 
 import (
 	"fmt"
+	"os"
+
+	"cosmossdk.io/client/v2/autocli"
+	clientv2keyring "cosmossdk.io/client/v2/autocli/keyring"
+	"cosmossdk.io/core/address"
+	"cosmossdk.io/core/appmodule"
+	"cosmossdk.io/depinject"
+	"cosmossdk.io/log"
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/config"
+	"github.com/cosmos/cosmos-sdk/codec"
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/cosmos/cosmos-sdk/x/auth/tx"
+	"github.com/cosmos/cosmos-sdk/x/auth/types"
+	pfm "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v8/packetforward"
+	pfmtypes "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v8/packetforward/types"
+	"github.com/cosmos/ibc-go/modules/capability"
+	capabilitytypes "github.com/cosmos/ibc-go/modules/capability/types"
+	ica "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts"
+	icatypes "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/types"
+	"github.com/cosmos/ibc-go/v8/modules/apps/transfer"
+	transfertypes "github.com/cosmos/ibc-go/v8/modules/apps/transfer/types"
+	ibc "github.com/cosmos/ibc-go/v8/modules/core"
+	ibcexported "github.com/cosmos/ibc-go/v8/modules/core/exported"
+	soloclient "github.com/cosmos/ibc-go/v8/modules/light-clients/06-solomachine"
+	tmclient "github.com/cosmos/ibc-go/v8/modules/light-clients/07-tendermint"
 	"github.com/noble-assets/noble/v8"
 	"github.com/noble-assets/noble/v8/cmd"
-	"os"
 )
 
 var (
@@ -24,6 +51,13 @@ var (
 	Bech32PrefixConsPub = Bech32PrefixAccAddr + "valconspub"
 )
 
+var (
+	txConfigOpts       tx.ConfigOptions
+	autoCliOpts        autocli.AppOptions
+	ModuleBasicManager module.BasicManager
+	clientCtx          client.Context
+)
+
 func main() {
 	config := sdk.GetConfig()
 	config.SetBech32PrefixForAccount(Bech32PrefixAccAddr, Bech32PrefixAccPub)
@@ -31,9 +65,76 @@ func main() {
 	config.SetBech32PrefixForConsensusNode(Bech32PrefixConsAddr, Bech32PrefixConsPub)
 	config.Seal()
 
-	rootCmd := cmd.NewRootCmd()
+	if err := depinject.Inject(
+		depinject.Configs(noble.AppConfig(),
+			depinject.Supply(
+				log.NewNopLogger(),
+			),
+			depinject.Provide(
+				ProvideClientContext,
+				ProvideKeyring,
+			),
+		),
+		&txConfigOpts,
+		&autoCliOpts,
+		&ModuleBasicManager,
+		&clientCtx,
+	); err != nil {
+		panic(err)
+	}
+
+	// Since the IBC modules don't support dependency injection, we need to
+	// manually register the modules on the client side.
+	// This needs to be removed after IBC supports App Wiring.
+	modules := map[string]appmodule.AppModule{
+		capabilitytypes.ModuleName: capability.AppModule{},
+		ibcexported.ModuleName:     ibc.AppModule{},
+		icatypes.ModuleName:        ica.AppModule{},
+		pfmtypes.ModuleName:        pfm.AppModule{},
+		transfertypes.ModuleName:   transfer.AppModule{},
+		tmclient.ModuleName:        tmclient.AppModule{},
+		soloclient.ModuleName:      soloclient.AppModule{},
+	}
+	for name, mod := range modules {
+		ModuleBasicManager[name] = module.CoreAppModuleBasicAdaptor(name, mod)
+		ModuleBasicManager[name].RegisterInterfaces(clientCtx.InterfaceRegistry)
+		autoCliOpts.Modules[name] = mod
+	}
+
+	rootCmd := cmd.NewRootCmd(txConfigOpts, autoCliOpts, ModuleBasicManager, clientCtx)
 	if err := svrcmd.Execute(rootCmd, "", noble.DefaultNodeHome); err != nil {
 		fmt.Fprintln(rootCmd.OutOrStderr(), err)
 		os.Exit(1)
 	}
+}
+
+func ProvideClientContext(
+	appCodec codec.Codec,
+	interfaceRegistry codectypes.InterfaceRegistry,
+	txConfig client.TxConfig,
+	legacyAmino *codec.LegacyAmino,
+) client.Context {
+	clientCtx := client.Context{}.
+		WithCodec(appCodec).
+		WithInterfaceRegistry(interfaceRegistry).
+		WithTxConfig(txConfig).
+		WithLegacyAmino(legacyAmino).
+		WithInput(os.Stdin).
+		WithAccountRetriever(types.AccountRetriever{}).
+		WithHomeDir(noble.DefaultNodeHome).
+		WithViper("") // env variable prefix
+
+	// Read the config again to overwrite the default values with the values from the config file
+	clientCtx, _ = config.ReadFromClientConfig(clientCtx)
+
+	return clientCtx
+}
+
+func ProvideKeyring(clientCtx client.Context, addressCodec address.Codec) (clientv2keyring.Keyring, error) {
+	kb, err := client.NewKeyringFromBackend(clientCtx, clientCtx.Keyring.Backend())
+	if err != nil {
+		return nil, err
+	}
+
+	return keyring.NewAutoCLIKeyring(kb)
 }
